Rewrote data (custom output) getting and setting functions.
Data is now stored in data attributes of the HTML elements that
comprise the choices for an input, and the content setter iterates
through the data array it receives to fill in the content for the ad.
This allows the playback mode (in which the page is passed an object of
values) to share the same code as the create/edit mode.